refactor: rename framebus file to framebus-config
