+++ spec/unit/methods/off.spec.ts
@@ -1,4 +1,4 @@
-import { FramebusConfig } from "../../../src/framebus";
+import { FramebusConfig } from "../../../src/framebus-config";
 import { subscribers } from "../../../src/lib/constants";
 import { off, on, teardown } from "../../../src/methods";
