+++ src/methods/include.ts
@@ -1,4 +1,4 @@
-import type { FramebusConfig } from "../framebus";
+import type { FramebusConfig } from "../framebus-config";
 import { childWindows } from "../lib/constants";
 
 export function include(config: FramebusConfig, childWindow: Window): boolean {
